requirements.txt
Flask==2.2.5
Flask[async]==2.2.5
boto3==1.34.56
python-telegram-bot==21.0.1
origamibot==2.3.6